backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
torch
torchvision
albumentations
opencv-python-headless
Pillow
numpy